### 2022-03-27 #522 @wfrierson
World Magnetic Model (WMM)
Submitting Author Name: Will Frierson
Submitting Author Github Handle: @wfrierson
Repository:  https://github.com/wfrierson/wmm
Version submitted: 1.1.1
Submission type: Standard
Version accepted: TBD
---

```
Package: wmm
Title: World Magnetic Model
Version: 1.1.1
Authors@R: person("Will", "Frierson", email = "will.frierson@gmail.com",
  role = c("aut", "cre"))
Maintainer: Will Frierson <will.frierson@gmail.com>
Description: Calculate magnetic field at a given location and time according to
  the World Magnetic Model (WMM). Both the main field and secular variation
  components are returned. This functionality is useful for physicists and
  geophysicists who need orthogonal components from WMM. Currently, this package
  supports annualized time inputs between 2000 and 2025. If desired, users can
  specify which WMM version to use, e.g., the original WMM2015 release or the
  recent out-of-cycle WMM2015 release. Methods used to implement WMM, including
  the Gauss coefficients for each release, are described in the following
  publications: Chulliat et al (2020) <doi:10.25923/ytk1-yx35>,
  Chulliat et al (2019) <doi:10.25921/xhr3-0t19>,
  Chulliat et al (2015) <doi:10.7289/V5TB14V7>,
  Maus et al (2010) <https://www.ngdc.noaa.gov/geomag/WMM/data/WMMReports/WMM2010_Report.pdf>,
  McLean et al (2004) <https://www.ngdc.noaa.gov/geomag/WMM/data/WMMReports/TRWMM_2005.pdf>,
  and Macmillian et al (2000) <https://www.ngdc.noaa.gov/geomag/WMM/data/WMMReports/wmm2000.pdf>.
  data.table (>= 1.12.2)
```

The `wmm` package retrieves magnetic field components given a location and time. I'm including "data munging" as recommended by @noamross ([link](https://github.com/ropensci/software-review/issues/461#issuecomment-915664660)).

Physicists, geophysicists, & navigators could find the magnetic field components useful.  Geo/physics students could find the construction of the magnetic field components helpful to their understanding.

To my knowledge, the most similar package is `oce` which has the function `magnetic.declination` that uses NOAA's WMM coefficients. However, the `oce` package does not provide magnetic field components, which are included in `wmm`.

n/a

Pre-submission inquiry: https://github.com/ropensci/software-review/issues/461

As noted in the pre-submission inquiry, I could not get `pkgcheck` to work. However, I read the guidelines associated with `pkgcheck` and I think `wmm` complies.
