### 2022-04-01 #522 @jooolia
Dear @wfrierson, thank you for your submission following your pre-submission.

Could you address the issues regarding the codemeta.json file, the contributing file, the vignette(s), and the examples. I see that you have continuous integration checks on Travis-CI so this X can be ignored. Please see the [packaging guide](https://devguide.ropensci.org/building.html) and feel free to ask any questions if you need help. Thanks, Julia
