### 2022-04-29 #522 @wfrierson
Thanks, Julia. Sorry for the delay! I'm working on these now.

Re: the examples, all the listed functions with a '.' prefix are not exported (and therefore not intended for end users) and the last 'wmm' reference is for the package-level wmm.R file, i.e., it's not a function. Do I still need examples for these?
