### 2022-05-01 #522 @jooolia
Hi @wfrierson,
Thanks for the update. Regarding the functions without examples, if you add the Roxygen  "@keywords internal" to these functions then this check should be ok for your "." fucntions.

Thanks, Julia
