### 2022-05-02 #522 @mpadge
One minor update on @jooolia's comment: the rOpenSci DevGuide [recommends using `@noRd` instead of `@Keywords internal`](https://devguide.ropensci.org/building.html#roxygen2-use).
