### 2022-06-03 #522 @wfrierson
Apologies for the delay. I'm updating my package on CRAN to be consistent
with the ropensci changes and am waiting for their approval. I'll check on
its status since it's been awhile.

- Will

On Fri, Jun 3, 2022, 7:28 AM Julia Gustavsen ***@***.***>
wrote:

> Hello @wfrierson <https://github.com/wfrierson>,
> Do you have any updates on your submission? Do you need any support from
> our side? Thanks, Julia
>
