### 2022-06-12 b15051a
ropensci_updates: Adding @noRd tags to internal functions.
