+++ R/functions_coefficients.R
@@ -2,6 +2,8 @@
 #'
 #' Find Gauss coefficient \eqn{g_{n,m}(t)} consistent with the World Magnetic Model.
 #'
+#' @noRd
+#'
 #' @param t Annualized date time. E.g., 2015-02-01 = (2015 + 32/365) = 2015.088
 #' @param t0 Annualized reference time associated with \code{t}
 #' @param wmmVersion String representing WMM version to use. Must be consistent with \code{time} and one of the following: 'derived', 'WMM2000', 'WMM2005', 'WMM2010', 'WMM2015', 'WMM2015v2', 'WMM2020'. Default 'derived' value will infer the latest WMM version consistent with \code{time}.
