### 2022-06-12 88a19c8
ropensci_updates: Incrementing package version, adding project status badge, and updating codemeta.json.
