### 2022-06-12 86a0302
ropensci_updates: Updating NEWS for v1.1.2.
