### 2022-06-12 #522 @wfrierson
@jooolia, is it possible for my latest changes to be reviewed before I submit to CRAN? If I submit to CRAN and then realize I missed something for ropensci, the CRAN volunteers will be less than excited to review a 2nd submission within the same month.

My changes to be compliant with ropensci are in the ropensci_updates branch ([link](https://github.com/wfrierson/wmm/tree/ropensci_updates)).
