### 2022-06-14 #522 @wfrierson
Ok, will do. Thanks!

On Tue, Jun 14, 2022, 11:38 AM Julia Gustavsen ***@***.***>
wrote:

> Hi @wfrierson <https://github.com/wfrierson>,
> I had a quick look and most of the issues appear to be fixed (I don't know
> why our CI checks are not being picked up, I will ask the other editors).
> Let me know when your changes will be merged into main and then I will find
> a handling editor who can then start looking for reviewers.
> Cheers, Julia
>
