### 2022-06-17 #522 @mpadge
@jooolia Our automated system only checks for continuous integration on GitHub. This package uses travis for CI, so that failing check may be ignored. (We [officially advise using alternatives to travis](https://devguide.ropensci.org/ci.html#travis-ci-linux-and-mac-osx), but that advise itself may need updating?)
