+++ man/dot-CalculateRadiusCurvature.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/functions_geocoordinates.R
-\name{.CalculateRadiusCurvature}
-\alias{.CalculateRadiusCurvature}
-\title{Radius of curvature of prime vertical}
-\usage{
-.CalculateRadiusCurvature(latitudeGD)
-}
-\arguments{
-\item{latitudeGD}{Geodetic latitude in decimal degrees}
-}
-\value{
-Radius of curvature of prime vertical at given geodetic latitude
-}
-\description{
-Calculate radius of curvature of prime vertical at given geodetic latitude.
-}
