+++ man/dot-DeriveVersionInfo.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/functions_misc.R
-\name{.DeriveVersionInfo}
-\alias{.DeriveVersionInfo}
-\title{Derive WMM version based on given time}
-\usage{
-.DeriveVersionInfo(t)
-}
-\arguments{
-\item{t}{Annualized date time. E.g., 2015-02-01 = (2015 + 32/365) = 2015.088}
-}
-\value{
-List of reference year and compatible WMM versions inferred from \code{time}.
-}
-\description{
-Derive WMM version based on given time
-}
