+++ man/dot-CalculateMagneticElements.Rd
@@ -1,17 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/functions_wmm.R
-\name{.CalculateMagneticElements}
-\alias{.CalculateMagneticElements}
-\title{Calculate Expected Magnetic Elements from WMM2020}
-\usage{
-.CalculateMagneticElements(orthComps)
-}
-\arguments{
-\item{orthComps}{named \code{list} containing magnetic orthogonal components}
-}
-\value{
-Expected magnetic components from WMM2020. \code{list}
-}
-\description{
-Calculate the magnetic elements (i.e., horizontal intensity, total intensity, inclination, declination, and their secular variation) for given magnetic orthogonal components
-}
