+++ man/dot-ConvertGeocentricToGeodeticFieldComponents.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/functions_geocoordinates.R
-\name{.ConvertGeocentricToGeodeticFieldComponents}
-\alias{.ConvertGeocentricToGeodeticFieldComponents}
-\title{Geocentric Coordinates to Geodetic Coordinates}
-\usage{
-.ConvertGeocentricToGeodeticFieldComponents(xGeocentric, yGeocentric,
-  zGeocentric, deltaLat)
-}
-\arguments{
-\item{xGeocentric}{X-coordinate in geocentric system}
-
-\item{yGeocentric}{Y-coordinate in geocentric system}
-
-\item{zGeocentric}{Z-coordinate in geocentric system}
-
-\item{deltaLat}{(Geocentric Latitude - Geodetic Latitude) in decimal degrees}
-}
-\value{
-Vector of length 3 representing geodetic coordinates consistent with given geocentric data
-}
-\description{
-Convert Geocentric Coordinates to Geodetic Coordinates.
-}
