### 2022-06-20 #522 @wfrierson
@jooolia, I just realized that I didn't update the NAMESPACE to reflect the @noRd changes. This is resolved in the latest commit. My apologies for that!

I see that ropensci-review-bot has noted the `renv` library. `wmm` is pure R, but I use `renv` for the packages used for unit tests. When installed from CRAN, no other packages are installed (noted in the DESCRIPTION file). Do I really need to remove `renv`?
