### 2022-06-20 #522 @mpadge
@wfrierson No, you don't need to *remove* `renv`, just [deactivate it](https://rstudio.github.io/renv/reference/deactivate.html). That only changes the local `.Rprofile` file by commenting out the `source` line. Alternatively, just remove that file from the repo and keep a local version only.
