### 2022-06-25 #522 @wfrierson
Thanks, @mpadge!

@jooolia, I've deactivated `renv`, which should be the last requested change. Are there any other changes needed?
