### 2022-06-27 #522 @jooolia
Hi @wfrierson,
To me it is looking good. I will look for a handling editor who can then look for reviewers.  Thanks!

@mpadge do you know why the internal functions are coming up as not having examples? This is why I suggested using "@keyword internal" because then for me the check was ok with pkgcheck(https://github.com/ropensci/software-review/issues/522#issuecomment-1114566751). Do you see the issue why it is not working with “@noRd“?
