### 2022-06-27 #522 @mpadge
@jooolia I've no idea, but nor is it repeatable: The bot checks on current system work fine and do not generate those failing "no examples" checks. Sorry for temporary inconvenience there.
