### 2022-06-29 #522 @maelle
:wave: @wfrierson!

- As noted by @mpadge above you should move away from Travis CI https://discuss.ropensci.org/t/ropensci-moving-away-from-travis-ci/2339/3 Therefore please switch to GitHub Actions or, say, Circle CI before we proceed. There are links in https://devguide.ropensci.org/ci.html#whichci (GitHub Actions is the most popular choice and well supported in the usethis package).
- It'd be nice to create a pkgdown website for reviewers to have this alternative to read the documentation. See https://devguide.ropensci.org/building.html#website
- A suggestion: could the README show a figure? (a graph created with numbers generated by using the package) I see the vignette is very short and does not show such an example.
- I made a small PR to rephrase a "click here" link https://github.com/wfrierson/wmm/pull/3
- Just a detail but you could make your GitHub profile more informative in particular pinning the package repository to it. https://devguide.ropensci.org/grooming.html#marketown

Happy to help with technical problems!
