### 2022-07-03 ee0c987
main: Adding github CI workflow.
