### 2022-07-03 c258e80
main: Removing travis CI.
