+++ .travis.yml
@@ -1,9 +1,0 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
-
-language: R
-cache: packages
-warnings_are_errors: false
-r_github_packages:
-  - r-lib/covr
-after_success:
-  - Rscript -e 'covr::coveralls()'
