### 2022-07-04 62c2565
main: Adding CRAN mirror to github actions yml.
