### 2022-07-09 5c45f92
main: Adding GITHUB_TOKEN to GH actions yml.
