+++ .github/workflows/r.yml
@@ -28,6 +28,8 @@ jobs:
           install.packages(c("remotes", "rcmdcheck"))
           remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
       - name: Check
         run: rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "error")
         shell: Rscript {0}
