+++ .github/workflows/r.yml
@@ -29,7 +29,7 @@ jobs:
           remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
         with:
-          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          repo-token: ${{ secrets.SECRET_PAT }}
       - name: Check
         run: rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "error")
         shell: Rscript {0}
