### 2022-07-09 68a8a67
main: Adding dependencies and unit tests to GH actions yml.
