### 2022-07-09 30f9da9
main: Adding pandoc to GH actions yml.
