### 2022-07-24 5a71d22
main: Ignoring pandoc in GH actions and adding coverage check.
