+++ .github/workflows/check_and_testthat.yml
@@ -18,7 +18,6 @@ jobs:
 
     steps:
       - uses: actions/checkout@v3
-      - uses: docker://pandoc/core:2.9
       - name: Set up R ${{ matrix.r-version }}
         uses: r-lib/actions/setup-r@f57f1301a053485946083d7a45022b278929a78a
         with:
@@ -30,7 +29,7 @@ jobs:
           remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
       - name: Check
-        run: rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "error")
+        run: rcmdcheck::rcmdcheck(args = "--no-manual --no-build-vignettes --ignore-vignettes", error_on = "error")
         shell: Rscript {0}
       - name: Run tests
         run: Rscript -e "source('tests/testthat.R')"
