### 2022-07-24 #522 @wfrierson
@maelle, I'm stumped on getting GH actions working. I've read tutorials and tried the `usethis` shortcuts. However, I continue to run into issue after issue. The script `check_and_testthat.yml` fails due to something about making the vignette into a pdf. The script `test-coverage.yaml` says it passes, but says "Request failed... Could not find a repository, try using repo upload token". I have researched these topics but with no progress.

I just need GH actions to do a few simple things:

- Run `check()`
- Check code coverage
- Run `testthat` scripts

If you have any recommendations, I'd appreciate any help ([link](https://github.com/wfrierson/wmm/tree/main/.github/workflows)). Thanks!
