### 2022-07-25 #522 @maelle
:wave: @wfrierson!

How did you create the `check_and_testthat.yml` file? I'd suggest deleting it, then updating usethis to its latest version, and then running `usethis::use_github_action_check_standard()` (that will create the R CMD check workflow). Then commit and push.

Regarding the test coverage error, do you have a codecov account linked to your GitHub account?
