### 2022-07-30 a4903a0
main: Switching to usethis template for GH actions R CMD check.
