+++ .github/workflows/check_and_testthat.yml
@@ -1,36 +1,0 @@
-name: R CI
-
-on:
-  push:
-    branches: [ "main" ]
-  pull_request:
-    branches: [ "main" ]
-
-permissions:
-  contents: read
-
-jobs:
-  build:
-    runs-on: macos-latest
-    strategy:
-      matrix:
-        r-version: ['3.6.3', '4.1.1']
-
-    steps:
-      - uses: actions/checkout@v3
-      - uses: r-lib/actions/setup-pandoc@v2
-      - name: Set up R ${{ matrix.r-version }}
-        uses: r-lib/actions/setup-r@f57f1301a053485946083d7a45022b278929a78a
-        with:
-          r-version: ${{ matrix.r-version }}
-      - name: Install dependencies
-        run: |
-          options(repos = c(CRAN = "https://cloud.r-project.org"))
-          install.packages(c("remotes", "rcmdcheck", "data.table", "knitr", "rmarkdown", "testthat"))
-          remotes::install_deps(dependencies = TRUE)
-        shell: Rscript {0}
-      - name: Check
-        run: rcmdcheck::rcmdcheck(args = "--no-manual --no-build-vignettes --ignore-vignettes", error_on = "error")
-        shell: Rscript {0}
-      - name: Run tests
-        run: Rscript -e "source('tests/testthat.R')"
