### 2022-07-30 #522 @wfrierson
Thanks for your help, @maelle! Sorry for being slow here. Thanks for your recommendation re: GH actions. I got the R CMD check workflow to work via `usethis`. I'm not troubleshooting running `testthat` unit tests.

@adamhsparks, do you happen to know of any example GH action workflows that demonstrate how to run a `testthat` sequence of unit tests?
