### 2022-08-03 #522 @adamhsparks
Hi @wfrierson, is this what you're after?

https://github.com/adamhsparks/wmm/runs/7644467088?check_suite_focus=true

I set this up using _usethis_ as below.

```r
# set up regular package checks and tests
usethis::use_github_action_check_standard()

# set up code coverage
usethis::use_coverage("codecov")

# automate the code coverage
usethis::use_github_action("test-coverage")
```
