### 2022-09-04 be53962
main: Applying usethis test-coverage template for GH coveralls action.
