+++ .github/workflows/unit-tests.yaml
@@ -1,41 +1,0 @@
-# Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
-# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
-on:
-  push:
-    branches: [main, master]
-  pull_request:
-    branches: [main, master]
-
-name: unit-tests
-
-jobs:
-  unit-tests:
-    runs-on: ubuntu-latest
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
-    steps:
-      - uses: actions/checkout@v2
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          use-public-rspm: true
-
-      - name: Install testthat
-        run: |
-          options(repos = c(CRAN = "https://cloud.r-project.org"))
-          install.packages("testthat")
-        shell: Rscript {0}
-
-      - name: Install data.table
-        run: |
-          options(repos = c(CRAN = "https://cloud.r-project.org"))
-          install.packages("data.table")
-        shell: Rscript {0}
-
-      - name: Run unit tests
-        run: |
-          library(testthat)
-          library(data.table)
-          testthat::test_check('wmm')
-        shell: Rscript {0}
