### 2022-09-04 #522 @wfrierson
@adamhsparks, your recommendation worked. Thank you!

@maelle, I've finally made progress with GH actions. Coverage is now tested and unit tests are checked! Hooray!

With this down, I can continue with your other requests. Thank you for your patience.
