### 2022-09-06 #522 @wfrierson
@maelle, can you help me understand this request?:

> A suggestion: could the README show a figure? (a graph created with numbers generated by using the package) I see the vignette is very short and does not show such an example.

The most natural graph for `wmm` is a 3D vector plot of the magnetic field components, but that doesn't seem practical as a text-only graph.
