### 2022-09-06 #522 @maelle
It is only a suggestion. If it is not natural to produce some sort of picture, then no need to! If you can create a plot then that'd be a nice illustration but really no obligation.
