### 2022-09-30 #522 @wfrierson
Hi @maelle,

If the plot is just a suggestion, I think it'd be better without it. The best picture that demonstrates the package is the vector plot found in the official NOAA report. Linking to the NOAA site seems adequate.

Are there any other issues that need to be resolved before you accept the package?
