### 2022-10-16 ee397f6
main: Adding rOpenSci peer review badge to README.
