### 2022-10-27 #522 @drmowinckels
I have never worked with the package authors

---

### Some note based on [rOpenSci packaging guidelines](https://devguide.ropensci.org/building.html), which are not mandatory but recommended. You _may_ consider adding:
- We recommend you to use the [codemetar package](https://github.com/ropensci/codemetar) for creating and updating a JSON [CodeMeta](https://codemeta.github.io/) metadata file for your package via `codemetar::write_codemeta()`
- If your package does not yet have a CITATION file, you can create one with [usethis::use_citation()](https://usethis.r-lib.org/reference/use_citation.html), and populate it with values generated by the [citation()](https://rdrr.io/r/utils/citation.html) function. [source](https://devguide.ropensci.org/building.html#citation-file)
- Consider using [usethis::use_readme_rmd()](https://usethis.r-lib.org/reference/use_readme_rmd.html) to get a template for a README.Rmd file and to automatically set up a pre-commit hook to ensure that README.md is always newer than README.Rmd.
- We recommend creating a documentation website for your package using [pkgdown](https://github.com/r-lib/pkgdown). Neal Richardson wrote [a good tutorial to get started with pkgdown](https://enpiar.com/2017/11/21/getting-down-with-pkgdown/), and unsurprisingly pkgdown has [its own documentation website](https://pkgdown.r-lib.org/).
- Consider adding contribution instructions and a CoC, so that people know how to help develop/bugfix or add extensions.

---

### Recommendation

- in the tests `testthat::context()` is [deprecated](https://testthat.r-lib.org/articles/third-edition.html#deprecations)
- you dont need to refer to testthat functions in the testfiles by `::`, they are all loaded by the testthat library call in testthat.R

Other than that, this is not my field, so I cannot say much about usage. But the package seems really nice in its simplicity, having a good coded backbone, with a single exported function for the end-user.  All examples, vignettes and code is working locally, and it was easy to test.

The code base uses camel-case, which is not my personal preference and not the recommendation by rOpenSci. But I dont think this is an issue, especially since the package exports a single function.
