### 2023-01-26 #522 @maelle
@HaoZeke thank you for your review!!

Regarding the code of conduct, after the reviews the rOpenSci code of conduct will apply to the package https://devguide.ropensci.org/collaboration.html?q=code%20of#coc-file (that same chapter contains tips on contributing guidelines)

Regarding argument validation, this might be an useful read: https://blog.r-hub.io/2022/03/10/input-checking/ (we have no specific guidelines on this in rOpenSci dev guide at the moment) (regarding strong typing @HaoZeke you might find https://rconsortium.github.io/OOP-WG/ and https://blog.q-lang.org/ interesting).
