### 2023-02-15 #522 @wfrierson
Thanks, @HaoZeke!

I'm looking into applying the recommended changes and will return when ready for review.
