### 2023-02-16 #522 @maelle
@wfrierson don't miss the [other review](https://github.com/ropensci/software-review/issues/522#issuecomment-1293953528) by @drmowinckels :wink:
