### 2023-05-02 #522 @maelle
@wfrierson any update? do we need to put this on hold?

From https://devguide.ropensci.org/policies.html#policiesreviewprocess

> The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
