### 2023-05-07 #522 @wfrierson
@HaoZeke, you commented:

> It would be nice to have more explicit contributing guidelines, and perhaps (but not necessarily required) a Code of conduct mention.

The package already includes a Code of Conduct section in `inst\docs\CONTRIBUTING.md`. Is this sufficient?
