### 2023-05-08 #522 @maelle
Why not store it in `.github/CONTRIBUTING.md` where it will be picked up by GitHub and pkgdown?
