### 2023-05-15 81d47fc
package_review: Moving CONTRIBUTING to .github.
