### 2023-05-15 #522 @wfrierson
@HaoZeke, no worries. I've moved it to .github.

Do you have any recommendations for this item?:
> adding a simple line about running the tests would be useful

The current CRAN version doesn't include the unit tests, but the GitHub project does. Is there a preferred way to do this?
