### 2023-05-23 #522 @maelle
Do you mean the tests are skipped on CRAN, and only run on GitHub Actions? This can make sense in some cases for sure, but you need to make sure tests are run regularly on GitHub Actions, for instance every week even when there's no change in the code.
