### 2023-06-02 e70f6fb
package_review: Updating vignette files for ropensci.
