### 2023-06-02 #522 @wfrierson
@maelle, I've setup GitHub Actions to run with new commits on the main branch. I've never heard of running tests when code has **not** been changed. Is this a standard practice elsewhere? If so, is implementing it as easy as changing my existing Actions to include something like the following?:

```
schedule:
  - cron: "* * * * *"
# Or whatever is needed for a weekly schedule...
```

I thought @HaoZeke was recommending a brief section in the README that shows how to run the tests. If that perspective is correct, I'm unclear how to do so. I haven't found examples or documentation that give a solution that's as easy as running:

```
devtools::test()
```

My apologies if I'm overthinking this. Thanks!
