### 2023-06-02 #522 @maelle
@wfrierson The reason why it might be useful to run tests on a weekly basis even in the absence of code changes is that you could catch problems with the upstream dependencies of your package.
Yes you simply need to add the cron field (but also keep the push one). Example https://github.com/hadley/r-pkgs/blob/4088c6e8d47c17138d8000fb299390c86fc74274/.github/workflows/render.yaml#L3 (the workflow dispatch one is useful, it's for having a button to launch the workflow from the actions tab).

@HaoZeke could you please clarify what you expected?
