### 2023-06-02 #522 @drmowinckels
@wfrierson I frequently use crontab guru for help in setting up correct cron specifications.
https://crontab.guru/every-sunday
