+++ .github/workflows/test-coverage.yaml
@@ -5,6 +5,9 @@ on:
     branches: [main, master]
   pull_request:
     branches: [main, master]
+  schedule:
+    # At 19:30 on Sunday
+    - cron: '30 19 * * sun'
 
 name: test-coverage
 
