### 2023-06-14 #522 @wfrierson
Thanks, @maelle & @drmowinckels!

I've added cron jobs for the two Actions. I'll merge to main after hearing back from @HaoZeke.
