### 2023-08-22 #522 @HaoZeke
Hi @wfrierson , @maelle and @drmowinckels . Yup the `cron` jobs are what I had in mind, though the readme addition would also be nice, since it currently isn't clear from the `readme` that tests are present.
