### 2023-09-05 #522 @wfrierson
Thanks, @HaoZeke. I think I've made all the requested changes. That said, I'm waiting to confirm that the `cron` job is working as intended.
