### 2023-09-06 #522 @wfrierson
@HaoZeke & @maelle, it looks like the `cron` job is working. Also, I updated the README to reference the unit tests.

Let me know if anything else is needed.
