### 2023-09-07 #522 @maelle
@wfrierson yes a bit more work! :grin: You'll need to write a response to the two reviews

And then please post your response with `@ropensci-review-bot submit response <url to issue comment>

Here's the [author guide for responding](https://devguide.ropensci.org/authors-guide.html)
