### 2023-09-25 #522 @wfrierson
@drmowinckels, in reviewing the comments above, I realized I forgot to respond to your recommendations. I'm so sorry!

Responses below:

- > in the tests testthat::context() is [deprecated](https://testthat.r-lib.org/articles/third-edition.html#deprecations)

Thanks for pointing that out. I wasn't aware. I've removed those.

- > you dont need to refer to testthat functions in the testfiles by ::, they are all loaded by the testthat library call in testthat.R

I've made this change in the latest commit.

- > consider adding another function to your package that is a simpler name for the main function. Its a long function name, while explanatory, a little much. I think having wmm::wmm() as a single function from your package would be quite elegant.

Yeah, it's name isn't great. My motivation: I wanted to make clear that this gives you the magnetic fields, since I couldn't find something like that before in R. Also, at this point, I feel stuck with it.
