### 2023-09-25 #522 @drmowinckels
Hi @wfrierson .

Thats all right. Thanks for addressing all my comments. I'd like to follow up on the last one. I think you can still have the original name, but also have a second "shorthand" function which just is a pointer to the same function.

An example of this is ggplot2's interchangebility of `colour` vs `color` in its scales.
https://github.com/tidyverse/ggplot2/blob/HEAD/R/zxx.R#L154

This being said, I'm by no means forcing this issue. If you want to keep it as is, that is also completely fine. I just wanted to point out the possibility.
