### 2023-10-07 56afdf6
main: Adding wmm alias to GetManeticFieldWMM.
