### 2023-10-07 8813008
main: Updating DESCRIPTION, NAMESPACE, and manual.
