+++ NAMESPACE
@@ -1,3 +1,4 @@
 # Generated by roxygen2: do not edit by hand
 
 export(GetMagneticFieldWMM)
+export(wmm)
