### 2023-10-08 #522 @wfrierson
@drmowinckels, I like your idea (and it's also a new thing I can learn in R). I've added the `wmm` alias for the main function in the latest commit. Thanks again for your feedback!

@maelle, unless there's anything else, I'll continue compiling my responses.
