### 2023-10-23 #522 @wfrierson
@maelle, thanks again for your patience and help! Please let me know if I've omitted anything and I can correct it.

Responses to reviewers are below.

### Response to @joolia:

- [x] Add codemeta.json
**RESPONSE**: Added in commit [6aa1dc2](https://github.com/wfrierson/wmm/blob/6aa1dc2cbce10f65b328f66b3979874fa6c8aa59/codemeta.json)
- [x] Add CONTRIBUTING
**RESPONSE**: Added in commit [eca5f5e](https://github.com/wfrierson/wmm/blob/eca5f5e960ba8d956a5bf0b4f908c90d13126840/inst/docs/CONTRIBUTING.md)
- [x] Add vignettes
**RESPONSE**: Added in commit [e70f6fb](https://github.com/wfrierson/wmm/tree/e70f6fb3a1b45c7364f111c2a532f530444c0057/vignettes)
- [x] Add @nord to non-exported functions
**RESPONSE**: Added in commit [b15051a](https://github.com/wfrierson/wmm/tree/b15051a06bf2e7a29a669ee87f6689141df5f821/R)

### Response to @mpadge
- [x] Deactivate ren in .Rprofile
**RESPONSE**: Updated in commit [6f5ecb5](https://github.com/wfrierson/wmm/blob/6f5ecb532db36d7ddddb8f0a8440bac10999eb28/.Rprofile)

### Response to @maelle
- [x] Switch from Travis CI to GitHub Actions
**RESPONSE**: This is complete and details can be found in the [Actions tab](https://github.com/wfrierson/wmm/actions)
- [x] Update WMM url reference
**RESPONSE**: Updated in commit [e306fb4](https://github.com/wfrierson/wmm/blob/e306fb4c99d7b19517eace3f224f297d33c461ee/README.md)

### Response to @HaoZeke
- [x] Include devtools:: based installation instructions in README
**RESPONSE**: Added in commit [9fb667c](https://github.com/wfrierson/wmm/blob/9fb667cec51023fef2d9f3482d6d66728c867588/README.md)
- [x] Include wmm:: reference in README examples
**RESPONSE**: Added in commit [9fb667c](https://github.com/wfrierson/wmm/blob/9fb667cec51023fef2d9f3482d6d66728c867588/README.md)
- [x] Add comment about running unit tests in README
**RESPONSE**: Added in commit [23aa014](https://github.com/wfrierson/wmm/blob/23aa0144b61c50b0a083fe0e52cd1eb61d9143cb/README.md)
- [x] Move CONTRIBUTING file to .github/CONTRIBUTING.md
**RESPONSE**: Updated in commit [81d47fc](https://github.com/wfrierson/wmm/blob/81d47fc220eb3114a0f698525c2eacc6c68c0cf8/.github/CONTRIBUTING.md)
- [x] Add weekly cron job for unit tests
**RESPONSE**: Added in commit [5af3489](https://github.com/wfrierson/wmm/tree/5af3489e30515e02705aa75111655f0d0110c693/.github/workflows)

Thank you to all my reviewers. You've helped identify many issues in this package which lead to its improvement. I can appreciate that you would've preferred to use your time in other ways; and so, you all were kind to contribute here.
