### 2023-10-30 #522 @maelle
@wfrierson thank you! Could you please collate your response to @drmowinckels in the same comment please so I can log it?

@drmowinckels @HaoZeke if you're satisfied with the author response, could you please use the [reviewer approval template](https://devdevguide.netlify.app/approval2template)? Thank you!
