### 2023-10-30 #522 @ropensci-review-bot
Approved! Thanks @wfrierson for submitting and @drmowinckels, @HaoZeke for your reviews! :grin:
