### 2023-11-04 #522 @wfrierson
Thanks, @maelle. I've added my response to @drmowinckels above. I also accepted the rOpenSci invitation.

Questions for you: will I have control of the repo after transferring to rOpenSci? I need to update it every 5 years to be current with the latest coefficients. Similarly, would anyone in rOpenSci make changes to the repo without my knowledge?

These are probably silly questions, but I've never transferred a repo before. Thanks for your patience.
