### 2023-11-09 #522 @maelle
>  Will I have control of the repo after transferring to rOpenSci? I need to update it every 5 years to be current with the latest coefficients.

Yes you will receive admin rights again after transfer.

> When I do update the coefficients, do I need to inform rOpenSci?

No.

> Similarly, would anyone in rOpenSci make changes to the repo without my knowledge?

Mainly, no, as you're the package maintainer. Small exceptions:

- We would add the repo URL to point to docs.ropensci.org if you don't do it and there's no other informative URL (on the right when looking at the repo).
- If say the pkgdown website isn't building because of a configuration problem, and you don't respond to a PR within months, then we might merge it ourselves to ensure the docs are still built.
We really don't want to step on anyone's toes, we respect your ownership.

See also https://devdevguide.netlify.app/maintenance_curation#peer-reviewed-packages

> These are probably silly questions, but I've never transferred a repo before. Thanks for your patience.

These are not silly! If you prefer, you can keep the repo under your own account. We'd still build the docs for docs.ropensci.org

Also relevant: https://devdevguide.netlify.app/maintenance_cheatsheet

Hope this helps!
