### 2023-11-23 #522 @wfrierson
Thanks again!

One more question: do you know if package submissions to CRAN need to be changed, once the package is in rOpenSci? I know this is more of a CRAN question, but I'm asking in case this has come up before. If it hasn't, no worries.
