### 2023-11-23 #522 @maelle
what do you mean by changing the package submissions to CRAN?

Don't worry about asking questions, we're here to help!
