### 2023-12-09 #522 @wfrierson
> what do you mean by changing the package submissions to CRAN?

I could imagine CRAN adding application steps for packages related to organizations, like rOpenSci. I don't think this exists, but instead wanted to confirm submitting to CRAN hadn't changed.

> If you prefer, you can keep the repo under your own account

Ok, I'll opt for keeping it on my account, out of simplicity. Thank you for being flexible on this!

---

With this approach in mind, are the following the only open items that remain (i.e., the two unchecked boxes below)?:

- _[n/a] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo. I have invited you to a team that should allow you to do so. You will need to [enable two-factor authentication](https://docs.github.com/en/authentication/securing-your-account-with-two-factor-authentication-2fa/configuring-two-factor-authentication) for your GitHub account.
This invitation will expire after one week. If it happens write a comment @ropensci-review-bot invite me to ropensci/<package-name> which will re-send an invitation._
- [n/a] _After transfer write a comment @ropensci-review-bot finalize transfer of <package-name> where <package-name> is the repo/package name. This will give you admin access back._
- _[n/a] Fix all links to the GitHub repo to point to the repo under the ropensci organization._
- [x] If you already had a pkgdown website and are ok relying only on [rOpenSci central docs building and branding](https://devguide.ropensci.org/ci.html#even-more-ci-ropensci-docs),
deactivate the automatic deployment you might have set up
remove styling tweaks from your pkgdown config but keep that config file
replace the whole current pkgdown website with a [redirecting page](https://devguide.ropensci.org/redirect.html)
replace your package docs URL with https://docs.ropensci.org/package_name
In addition, in your DESCRIPTION file, include the docs link in the URL field alongside the link to the GitHub repository, e.g.: URL: https://docs.ropensci.org/foobar, https://github.com/ropensci/foobar
 **Increment the package version to reflect the changes you made during review. In NEWS.md, add a heading for the new version and one bullet for each user-facing change, and each developer-facing change that you think is relevant.**
- [ ] You can add this installation method to your package README install.packages("<package-name>", repos = "https://ropensci.r-universe.dev") thanks to [R-universe](https://ropensci.org/blog/2021/06/22/setup-runiverse/).
