### 2023-12-12 #522 @maelle
http://docs.ropensci.org/wmm/ is up :tada: you can store this URL in the DESCRIPTION file as well as the GitHub repository description.

Link to our maintainer cheatsheet: https://devdevguide.netlify.app/maintenance_cheatsheet
